refactor: common server helpers
